move out file protocol link resolution test as it does not make sense for JS
