+++ src/jvmTest/resources/data/html/baseUriFile.txt
